+++ problems/sedov/initproblem.F90
@@ -38,8 +38,8 @@ module initproblem
    private
    public  :: read_problem_par, init_prob, problem_pointers
 
-   integer :: n_sn
-   real    :: d0, p0, bx0, by0, bz0, Eexpl, x0, y0, z0, r0, dt_sn, r, t_sn
+   integer(kind=4) :: n_sn
+   real            :: d0, p0, bx0, by0, bz0, Eexpl, x0, y0, z0, r0, dt_sn, r, t_sn
 
    namelist /PROBLEM_CONTROL/ d0, p0, bx0, by0, bz0, Eexpl, x0, y0, z0, r0, n_sn, dt_sn
